+++ src/com/firstApp/Cliente.java
@@ -35,4 +35,52 @@ public class Cliente extends Pessoa {
     public void setCredito(float credito) {
         this.credito = credito;
     }
+
+    public static class Endereço {
+       private String rua;
+       private String bairro;
+       private String estado;
+       private String regiao;
+       private int numero;
+
+       public String getRua() {
+           return rua;
+       }
+
+       public void setRua(String rua) {
+           this.rua = rua;
+       }
+
+       public String getBairro() {
+           return bairro;
+       }
+
+       public void setBairro(String bairro) {
+           this.bairro = bairro;
+       }
+
+       public String getEstado() {
+           return estado;
+       }
+
+       public void setEstado(String estado) {
+           this.estado = estado;
+       }
+
+       public String getRegiao() {
+           return regiao;
+       }
+
+       public void setRegiao(String regiao) {
+           this.regiao = regiao;
+       }
+
+       public int getNumero() {
+           return numero;
+       }
+
+       public void setNumero(int numero) {
+           this.numero = numero;
+       }
+   }
 }
