+++ src/com/firstApp/Main.java
@@ -4,6 +4,7 @@ public class Main {
 
     public static void main(String[] args) {
 	// write your code here
-    
+        Cliente cliente = new Cliente("Tiago",18,1101010);
+        System.out.println(cliente.toString());
     }
 }
